backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
uvicorn[standard]==0.23.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==1.7.0
python-dotenv==1.0.0
pydantic==2.5.0
ollama==0.1.7
openai==1.3.5
anthropic==0.3.15
pytest==7.4.3
pytest-asyncio==0.21.1
mcp==0.1.0